% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalQual.R
\name{causalQual}
\alias{causalQual}
\title{Causal Inference with Qualitative Outcomes}
\usage{
causalQual(
  Y = NULL,
  Y_pre = NULL,
  Y_post = NULL,
  D = NULL,
  X = NULL,
  Z = NULL,
  running_variable = NULL,
  cutoff = NULL,
  identification = NULL,
  outcome_type = NULL,
  K = 5
)
}
\arguments{
\item{Y}{Qualitative outcome. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{Y_pre}{Qualitative outcome before treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{Y_post}{Qualitative outcome after treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{D}{Binary treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{Z}{Binary instrument.}

\item{running_variable}{Running variable determining treatment assignment.}

\item{cutoff}{Cutoff or threshold. Units with \code{running_variable < cutoff} are considered controls, while units with \code{running_variable > cutoff} are considered treated.}

\item{identification}{String controlling the employed identification strategy. Must be one of \code{"selection_on_observables"}, \code{"iv"}, \code{"diff_in_diff"}, \code{"rd"}.}

\item{outcome_type}{String controlling the outcome type. Must be either \code{"multinomial"} or \code{"ordered"}. Affects estimation of conditional class probabilities.}

\item{K}{Number of folds for nuisance functions estimation.}
}
\value{
An object of class \code{\link{causalQual}}.
}
\description{
Estimates treatment effects on qualitative outcomes (multinomial or ordered non-numeric) under several research designs.
}
\details{
Not all arguments must be used when calling \code{\link{causalQual}}. Specifically, the required arguments depend on the choice of \code{"identification"}.
If unnecessary arguments are used, \code{\link{causalQual}} raises an error.
\subsection{Selection-on-observables}{

Under a selection-on-observables design, identification requires the treatment indicator to be (conditionally) independent of potential outcomes
(unconfoundedness), and that each unit has a non-zero probability of being treated (common support). If these assumptions hold, we can recover the
probabilities for shift of all classes.\cr

If \code{identification == "selection_on_observables"}, \code{\link{causalQual}} constructs and averages doubly robust scores for qualitative outcomes
to estimate the probability shift \eqn{\delta_m := P(Y(1) = m) - P(Y(0) = m)}. For each class \eqn{m}, the doubly robust score is defined as:
\deqn{
    \hat{\Gamma}_{m, i} =
    \hat{P}(Y_i = m \mid D_i = 1, X_i) - \hat{P}(Y_i = m \mid D_i = 0, X_i) +
    D_i \frac{1\{Y_i = m\} - \hat{P}(Y_i = m \mid D_i = 1, X_i)}{\hat{P}(D_i = 1 | X_i)}
    - (1 - D_i) \frac{1\{Y_i = m\} - \hat{P}(Y_i = m \mid D _i= 0, X_i)}{1 - \hat{P}(D_i = 1 | X_i)}.
}

The estimator for the probability shift of class \eqn{m} is then the average of the scores:
\deqn{
    \hat{\delta}_m = \frac{1}{n} \sum_{i=1}^{n} \hat{\Gamma}_{m, i}.
}

with its variance estimated as:
\deqn{
    \widehat{\text{Var}} ( \hat{\delta}_m ) = \frac{1}{n} \sum_{i=1}^{n} ( \hat{\Gamma}_{m, i} - \hat{\delta}_m )^2.
}

\code{\link{causalQual}} uses these estimates to construct confidence intervals based on conventional normal approximations.

If \code{outcome_type == "multinomial"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using a \code{\link[ocf]{multinomial_ml}} strategy with regression forests
as base learners. Else, if \code{outcome_type == "ordered"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using the honest version of the \code{\link[ocf]{ocf}} estimator.
\eqn{\hat{P}(D_i = 1 | X_i)} is always estimated via a honest \code{\link[grf]{regression_forest}}. \code{K}-fold cross-fitting is employed for the estimation of all these functions.
}

\subsection{Instrumental-variables}{

Under an instrumental-variables design, identification requires the instrument to be independent of potential outcomes and potential treatments (exogeneity), that the
instrument influences the outcome solely through its effect on treatment (exclusion restriction), that the instrument has a nonzero effect on treatment probability (relevance), and that the instrument can only
increase/decrease the treatment probability (monotonicity). If these assumptions hold, we can recover the local probabilities of shift for all classes.\cr

If \code{identification == "iv"}, \code{\link{causalQual}} applies, for each class \code{m}, the standard two-stage least squares method to the binary variable \eqn{1(Y_i = m)}. Specifically, the routine first estimates
the following first-stage regression model via OLS:

\deqn{D_i = \gamma_0 + \gamma_1 Z_i + \nu_i,}

and constructs the predicted values \eqn{\hat{D}_i}. It then uses these predicted values in the second-stage regression:

\deqn{1(Y_i = m) = \alpha_{m0} + \alpha_{m1} \hat{D}_i + \epsilon_{mi}.}

The OLS estimate \eqn{\hat{\alpha}_{m1}} of \eqn{\alpha_{m1}} is then our estimate of the local probability shift for class \code{m}. Standard errors are computed using conventional procedures and used to construct
conventional confidence intervals. All of this is done by calling the \code{\link[AER]{ivreg}} function.
}

\subsection{Regression discontinuity}{

Under a regression discontinuity design, identification requires that the probability mass functions for class \eqn{m} of potential outcomes are continuous in the running variable (continuity). If this assumption holds,
we can recover the probability shift at the cutoff for class \eqn{m}.\cr

If \code{identification == "rd"}, \code{\link{causalQual}} applies, for each class \eqn{m}, standard local polynomial estimators to the binary variable \eqn{1(Y_i = m)}. Specifically, the ruotine implements the
robust bias-corrected inference procedure of Calonico et al. (2014) (see the \code{\link[rdrobust]{rdrobust}} function).
}

\subsection{Difference-in-differences}{

Under a difference-in-difference design, identification requires that the probabilities time shift for class \eqn{m} evolve similarly for the treated and control groups (parallel
trends on the probability mass functions of \eqn{Y(0)}). If this assumption holds, we can recover the probability of shift on the treated for class \eqn{m}.\cr

If \code{identification == "diff_in_diff"}, \code{\link{causalQual}} applies, for each class \eqn{m}, the canonical two-group/two-period method to the binary variable \eqn{1(Y_{is} = m)}. Specifically,
consider the following linear model:

\deqn{1(Y_{is} = m) = D_i \beta_{m1} + 1(s = t) \beta_{m2} + D_i 1(s = t) \beta_{m3} + \epsilon_mis.}

The OLS estimate \eqn{\hat{\beta}_{m3}} of \eqn{\beta_{m3}} is our estimate of the probability shift on the treated for class \code{m}. Standard errors are clustered at the unit level and used to construct
conventional confidence intervals.
}
}
\examples{
\donttest{## Selection-on-observables.
# Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(1000, assignment = "observational",
                                      outcome_type = "ordered")
Y <- data$Y
D <- data$D
X <- data$X

# Estimate probabilities of shift.
fit <- causalQual(Y = Y, D = D, X = X,
                  identification = "selection_on_observables",
                  outcome_type = "ordered", K = 2)

summary(fit)
plot(fit)

## Instrumental variables.
# Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_iv(1000,
                                     outcome_type = "ordered")
Y <- data$Y
D <- data$D
Z <- data$Z

# Estimate local probabilities of shift.
fit <- causalQual(Y = Y, D = D, Z = Z,
                  identification = "iv",
                  outcome_type = "ordered")

summary(fit)
plot(fit)

## Regression discontinuity.
# Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_rd(1000,
                                     outcome_type = "ordered")
Y <- data$Y
running_variable <- data$running_variable
cutoff <- data$cutoff

# Estimate probabilities of shift at the cutoff.
fit <- causalQual(Y = Y, running_variable = running_variable, cutoff = cutoff,
                  identification = "rd")

summary(fit)
plot(fit)

## Difference-in-differences.
# Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_did(1000, assignment = "observational",
                                      outcome_type = "ordered")
Y_pre <- data$Y_pre
Y_post <- data$Y_post
D <- data$D

# Estimate probabilities of shift on the treated.
fit <- causalQual(Y_pre = Y_pre, Y_post = Y_post, D = D,
                  identification = "diff_in_diff",
                  outcome_type = "ordered")

summary(fit)
plot(fit)}

}
\author{
Riccardo Di Francesco
}
