% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{causalQual_did}
\alias{causalQual_did}
\title{Two-Way Fixed Effects Models for Qualitative Outcomes}
\usage{
causalQual_did(Y_pre, Y_post, D)
}
\arguments{
\item{Y_pre}{Qualitative outcome before treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{Y_post}{Qualitative outcome after treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{D}{Binary treatment indicator.}
}
\value{
A list with estimates, standard errors, and model fit information.
}
\description{
Fit two-way fixed-effect models for qualitative outcomes to estimate the probabilities of shift on the treated. Identification requires that the probability time shifts follow
a similar evolution over time in both the treated and control groups (parallel trends on the mass functions).
}
\details{
This function fits one linear model for each class \eqn{m} of \code{Y}. Each model uses a binary indicator \code{Y == m} as an outcome and a time dummy, a treatment dummy (\code{D}), and an interaction thereof
as regressors. Standard errors are clustered at the unit level.
}
\author{
Your Name
}
