% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{causalQual_rd}
\alias{causalQual_rd}
\title{Local Polynomial Regression for Qualitative Outcomes}
\usage{
causalQual_rd(Y, running_variable, cutoff)
}
\arguments{
\item{Y}{Qualitative outcome before treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{running_variable}{Running variable determining treatment assignment.}

\item{cutoff}{Cutoff or threshold. Units with \code{running_variable < cutoff} are considered controls, while units with \code{running_variable > cutoff} are considered treated.}
}
\value{
A list with estimates, standard errors, and model fit information.
}
\description{
Local polynomial regression models for qualitative outcomes to estimate the probabilities of shift at the cutoff. Identification requires that the probability mass functions of potential outcomes
are continuous in the running variable (continuity).
}
\details{
This function fits local linear regressions to either side of the cutoff using optimal bandwidth calculation.
It then constructs robust bias-corrected confidence intervals. All of this is done by calling the \code{\link[rdrobust]{rdrobust}} function.
}
\author{
Riccardo Di Francesco
}
