% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_qualitative_data_did}
\alias{generate_qualitative_data_did}
\title{Generate Qualitative Data (Difference-in-Differences)}
\usage{
generate_qualitative_data_did(n, assignment, outcome_type)
}
\arguments{
\item{n}{Sample size.}

\item{assignment}{String controlling treatment assignment. Must be either \code{"randomized"} (random assignment)
or \code{"observational"} (assignment based on covariates).}

\item{outcome_type}{String controlling the outcome type. Must be either \code{"multinomial"} or \code{"ordered"}.}
}
\value{
A list storing a data frame with the observed data, the true propensity score, and the true probabilities of shift on the treated.
}
\description{
Generate a synthetic data set with qualitative outcomes for a difference-in-differences setting. The data include two time periods, a binary treatment indicator (applied only in the second period),
and a matrix of covariates. Probabilities time shift among the treated and control groups evolve similarly across the two time periods (parallel trends on the probability mass functions).\cr\cr
\strong{WARNING} In the current implementation, parallel trends does not hold if \code{outcome_type == "ordered"} and \code{assignment == "observational"}.
}
\details{
\subsection{Outcome type}{

Potential outcomes are generated differently according to \code{outcome_type}. If \code{outcome_type == "multinomial"}, the function computes linear predictors for each class using the covariates:

\deqn{\eta_{mi} (d, s) = \gamma 1(s = t) + \beta_{m1}^d X_{i1} + \beta_{m2}^d X_{i2} + \beta_{m3}^d X_{i3},}

and then transforms \eqn{\eta_{mi} (d, s)} into valid probability distributions using the softmax function:

\deqn{P(Y(d, s) = m \mid X) = \frac{\exp(\eta_{mi (d, s)})}{\sum_{m'} \exp(\eta_{m'i(d, s)})}.}

It then generates potential outcomes \eqn{Y(1, t-1)}, \eqn{Y(1, t)}, \eqn{Y(0, t-1)}, and \eqn{Y(0, t)} by sampling from \{1, 2, 3\} using \eqn{P(Y(d, s) = m \mid X)}.\cr

If instead \code{outcome_type == "ordered"}, we first generate latent potential outcomes:

\deqn{Y_i^* (d, s) = \tau d + \gamma 1(s = t) + X_{i1} + X_{i2} + X_{i3} + N (0, 1),}

with \eqn{\tau = 2}. We then construct \eqn{Y_i (d, s)} by discretizing \eqn{Y_i^* (d, s)} using threshold parameters \eqn{\zeta_1 = 2} and \eqn{\zeta_2 = 4}. We then have

\deqn{P(Y_i(d, s) = m) = P(\zeta_{m-1} < Y_i^*(d, s) \leq \zeta_m) = \Phi (\zeta_m - \sum_j X_{ij} - \tau d - \gamma 1(s = t)) - \Phi (\zeta_{m-1} - \sum_j X_{ij} - \tau d - \gamma 1(s = t)).}

which allows us to analytically compute the probabilities of shift.
}

\subsection{Treatment assignment}{

Treatment is always assigned as \eqn{D_i \sim \text{Bernoulli}(\pi(X_i))}. If \code{assignment == "randomized"}, then the propensity score is specified as \eqn{\pi(X_i) = P ( D_i = 1 | X_i)) = (X_{i1} + X_{i2}) / 2}.
If instead \code{assignment == "observational"}, then \eqn{\pi(X_i) = 0.5}.
}

\subsection{Other details}{

The function always generates three independent covariates from \eqn{U(0,1)}. Observed outcomes \eqn{Y_{is}} are always constructed using the usual observational rule.
}
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_did(100,
                                      assignment = "observational",
                                      outcome_type = "ordered")

data$pshifts_treated}

}
