% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.causalQual}
\alias{print.causalQual}
\title{Print Method for causalQual Objects}
\usage{
\method{print}{causalQual}(x, ...)
}
\arguments{
\item{x}{An \code{causalQual} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints an \code{causalQual} object.
}
\description{
Prints an \code{causalQual} object.
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(1000, assignment = "observational")
Y <- data$Y
D <- data$D
X <- data$X

## Estimate probability shifts.
fit <- causalQual(Y, D, X, 
                  identification = "selection_on_observables",
                  outcome_type = "ordered")
print(fit)}

}
\seealso{
\code{\link{causalQual}}
}
\author{
Riccardo Di Francesco
}
