% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softmax}
\alias{softmax}
\title{Softmax function.}
\usage{
softmax(logits)
}
\arguments{
\item{logits}{An nxM matrix, with n the sample size and M the number of classes of an ordinal outcome.}
}
\value{
The same matrix where each row is normalized so that row sums equal 1; useful to construct conditional probability mass functions.
}
\description{
Implementation of the softmax function.
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
