% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalQual.R
\name{causalQual}
\alias{causalQual}
\title{Causal Inference with Qualitative Outcomes}
\usage{
causalQual(Y, D, X, identification, outcome_type, K = 5)
}
\arguments{
\item{Y}{Ordered non-numeric outcome. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{D}{Binary treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{identification}{String controlling the employed identification strategy. Must be one of \code{"selection_on_observables"}.}

\item{outcome_type}{String controlling the outcome type. Must be either \code{"multinomial"} or \code{"ordered"}. Affects estimation of conditional class probabilities.}

\item{K}{Number of folds for nuisance functions estimation. Ignored if \code{identification != "selection_on_observables"}.}
}
\value{
An object of class \code{\link{causalQual}}.
}
\description{
Estimates treatment effects on qualitative outcomes (e.g., multinomial or ordered non-numeric) under several research designs.
}
\details{
Depending on \code{identification}, \code{\link{causalQual}} calls the appropriate estimation function.
\subsection{Selection-on-observables}{

Under selection-on-observables, identification requires the treatment indicator to be (conditionally) independent of potential outcomes
(unconfoundedness), and that each unit has a non-zero probability of being treated (common support).\cr

If \code{identification == "selection_on_observables"}, \code{\link{causalQual}} constructs and averages doubly robust scores for ordinal outcomes
to estimate the probability shifts \eqn{\delta_m := P(Y(1) = m) - P(Y(0) = m)}. For each class \eqn{m}, the doubly robust score is defined as:
\deqn{
    \hat{\Gamma}_{m, i} =
    \hat{P}(Y = m \mid D = 1, X_i) - \hat{P}(Y = m \mid D = 0, X_i) +
    D_i \frac{1\{Y = m\} - \hat{P}(Y = m \mid D = 1, X_i)}{\hat{P}(D_i = 1 | X_i)}
    - (1 - D_i) \frac{1\{Y = m\} - \hat{P}(Y = m \mid D = 0, X_i)}{1 - \hat{P}(D_i = 1 | X_i)}.
}

The estimator for the probability shift is then the average of the scores:
\deqn{
    \hat{\delta}_m = \frac{1}{n} \sum_{i=1}^{n} \hat{\Gamma}_{m, i}.
}

The variance of \eqn{\hat{\delta}_m} is estimated as:
\deqn{
    \widehat{\text{Var}} ( \hat{\delta}_m ) = \frac{1}{n} \sum_{i=1}^{n} ( \hat{\Gamma}_{m, i} - \hat{\delta}_m )^2.
}

If \code{outcome_type == "multinomial"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using a \code{\link[ocf]{multinomial_ml}} strategy with regression forests
as base learner. Else, if \code{outcome_type == "ordered"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using the honest version of the \code{\link[ocf]{ocf}} estimator.
\eqn{\hat{P}(D_i = 1 | X_i)} is always estimated via a honest \code{\link[grf]{regression_forest}}. K-fold cross-fitting is employed.
}
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(100, assignment = "observational")
Y <- data$Y
D <- data$D
X <- data$X

## Estimate probability shifts.
fit <- causalQual(Y, D, X, 
                  identification = "selection_on_observables",
                  outcome_type = "ordered")

summary(fit)
plot(fit)}
  
}
\seealso{
\code{\link{causalQual_soo}}
}
\author{
Riccardo Di Francesco
}
