% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{causalQual_iv}
\alias{causalQual_iv}
\title{Two-Stage Least Squares for Qualitative Outcomes}
\usage{
causalQual_iv(Y, D, Z)
}
\arguments{
\item{Y}{Qualitative outcome before treatment. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{D}{Binary treatment indicator.}

\item{Z}{Binary instrument.}
}
\value{
A list with estimates, standard errors, and model fit information.
}
\description{
Fit two-stage least squares models for qualitative outcomes to estimate the local probabilities of shift. Identification requires the instrument to be independent of potential outcomes and
potential treatments (exogeneity) and not to have any direct effect on potential outcomes (exclusion restriction). The instrument must also have a non-zero impact on treatment probability
(relevance) and to only increase/decrease that probability (monotonicity).
}
\details{
This function fits one two-stage least squares model for each class \eqn{m} of \code{Y}. Each model uses a binary indicator \code{Y == m} as an outcome.
}
\author{
Riccardo Di Francesco
}
