% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-functions.R
\name{causalQual_soo}
\alias{causalQual_soo}
\title{Double Machine Learning for Qualitative Outcomes}
\usage{
causalQual_soo(Y, D, X, outcome_type, K = 5)
}
\arguments{
\item{Y}{Ordered non-numeric outcome. Must be labeled as \eqn{\{1, 2, \dots\}}.}

\item{D}{Binary treatment indicator.}

\item{X}{Covariate matrix (no intercept).}

\item{outcome_type}{String controlling the outcome type. Must be either \code{"multinomial"} or \code{"ordered"}. Affects estimation of conditional class probabilities.}

\item{K}{Number of folds for nuisance functions estimation.}
}
\value{
An object of class \code{causalQual}.
}
\description{
Construct and average doubly robust scores for qualitative outcomes to estimate the probability shifts. Identification requires the treatment indicator to be (conditionally) independent of potential outcomes
(unconfoundedness), and that each unit has a non-zero probability of being treated (common support).
}
\details{
For each class \eqn{m}, the doubly robust score is defined as:
\deqn{
    \hat{\Gamma}_{m, i} =
    \hat{P}(Y = m \mid D = 1, X_i) - \hat{P}(Y = m \mid D = 0, X_i) +
    D_i \frac{1\{Y = m\} - \hat{P}(Y = m \mid D = 1, X_i)}{\hat{P}(D_i = 1 | X_i)}
    - (1 - D_i) \frac{1\{Y = m\} - \hat{P}(Y = m \mid D = 0, X_i)}{1 - \hat{P}(D_i = 1 | X_i)}.
}

The estimator for the probability shift is then the average of the scores:
\deqn{
    \hat{\delta}_m = \frac{1}{n} \sum_{i=1}^{n} \hat{\Gamma}_{m, i}.
}

The variance of \eqn{\hat{\delta}_m} is estimated as:
\deqn{
    \widehat{\text{Var}} ( \hat{\delta}_m ) = \frac{1}{n} \sum_{i=1}^{n} ( \hat{\Gamma}_{m, i} - \hat{\delta}_m )^2.
}

If \code{outcome_type == "multinomial"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using a \code{\link[ocf]{multinomial_ml}} strategy with regression forests
as base learner. Else, if \code{outcome_type == "ordered"}, \eqn{\hat{P}(Y = m \mid D = 1, X)} and \eqn{\hat{P}(Y = m \mid D = 0, X)} are estimated using the honest version of the \code{\link[ocf]{ocf}} estimator.
\eqn{\hat{P}(D_i = 1 | X_i)} is always estimated via a honest \code{\link[grf]{regression_forest}}. K-fold cross-fitting is employed.
}
\seealso{
\code{\link{causalQual}}
}
\author{
Riccardo Di Francesco
}
\keyword{internal}
