% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_qualitative_data_soo}
\alias{generate_qualitative_data_soo}
\title{Generate Qualitative Data (Selection-on-Observables)}
\usage{
generate_qualitative_data_soo(n, assignment, outcome_type)
}
\arguments{
\item{n}{Sample size.}

\item{assignment}{String controlling treatment assignment. Must be either \code{"randomized"} (random assignment) or \code{"observational"} (random assigment conditional on the generated covariates).}

\item{outcome_type}{String controlling the outcome type. Must be either \code{"multinomial"} or \code{"ordered"}. Affects how potential outcomes are constructed.}
}
\value{
A list storing a data frame with the observed data, the true propensity score, and the true probabilities of shift.
}
\description{
Generate a synthetic data set with a qualitative outcome, a binary treatment indicator, and a matrix of covariates. The treatment is either independent or conditionally independent of potential outcomes,
depending on users' choices.
}
\details{
\subsection{Outcome type}{

Potential outcomes are generated differently according to \code{outcome_type}. If \code{outcome_type == "multinomial"}, the function computes linear predictors for each class using the covariates:

\deqn{\eta_{mi} (d) = \beta_{m1}^d X_{i1} + \beta_{m2}^d X_{i2} + \beta_{m3}^d X_{i3},}

and then transforms \eqn{\eta_{mi} (d)} into valid probability distributions using the softmax function:

\deqn{P(Y(d) = m \mid X) = \frac{\exp(\eta_{mi (d)})}{\sum_{m'} \exp(\eta_{m'i(d)})}.}

It then generates potential outcomes \eqn{Y(1)} and \eqn{Y(0)} by sampling from \{1, 2, 3\} using \eqn{P(Y(d) = m \mid X)}.\cr

If instead \code{outcome_type == "ordered"}, we first generate latent potential outcomes:

\deqn{Y_i^* (d) = \tau d + X_{i1} + X_{i2} + X_{i3} + N (0, 1),}

with \eqn{\tau = 2}. We then construct \eqn{Y_i (d)} by discretizing \eqn{Y_i^* (d)} using threshold parameters \eqn{\zeta_1 = 2} and \eqn{\zeta_2 = 4}. We then have

\deqn{P(Y_i(d) = m) = P(\zeta_{m-1} < Y_i^*(d) \leq \zeta_m) = \Phi (\zeta_m - \sum_j X_{ij} - \tau d) - \Phi (\zeta_{m-1} - \sum_j X_{ij} - \tau d).}

which allows us to analytically compute the probabilities of shift.
}

\subsection{Treatment assignment}{

Treatment is always assigned as \eqn{D_i \sim \text{Bernoulli}(\pi(X_i))}. If \code{assignment == "randomized"}, then the propensity score is specified as \eqn{P ( D_i = 1 | X_i)) = (X_{i1} + X_{i2}) / 2}.
If instead \code{assignment == "observational"}, then \eqn{P ( D_i = 1 | X_i)) = 0.5}.
}

\subsection{Other details}{

The function always generates three independent covariates from \eqn{U(0,1)}. Observed outcomes \eqn{Y_i} are always constructed using the usual observational rule.
}
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(100,
                                      assignment = "observational",
                                      outcome_type = "ordered")

data$pshifts}

}
\seealso{
\code{\link{causalQual}}
}
\author{
Riccardo Di Francesco
}
