% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{plot.causalQual}
\alias{plot.causalQual}
\title{Plot Method for causalQual Objects}
\usage{
\method{plot}{causalQual}(x, hline = TRUE, ...)
}
\arguments{
\item{x}{An \code{\link{causalQual}} object.}

\item{hline}{Logical, whether to display an horizontal line at zero in the plot.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Plots an \code{\link{causalQual}} object.

\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(1000, assignment = "observational",
                                      outcome_type = "ordered")
Y <- data$Y
D <- data$D
X <- data$X

## Estimate probabilities of shifts.
fit <- causalQual(Y = Y, D = D, X = X,
                  identification = "selection_on_observables",
                  outcome_type = "ordered")
plot(fit)}
}
\description{
Plots an \code{\link{causalQual}} object.
}
\seealso{
\code{\link{causalQual}}
}
\author{
Riccardo Di Francesco
}
