% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.causalQual}
\alias{summary.causalQual}
\title{Summary Method for causalQual Objects}
\usage{
\method{summary}{causalQual}(object, ...)
}
\arguments{
\item{object}{An \code{\link{causalQual}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Summarizes an \code{\link{causalQual}} object.
}
\description{
Summarizes an \code{\link{causalQual}} object.
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_qualitative_data_soo(1000, assignment = "observational")
Y <- data$Y
D <- data$D
X <- data$X

## Estimate probability shifts.
fit <- causalQual(Y, D, X, 
                  identification = "selection_on_observables",
                  outcome_type = "ordered")
summary(fit)}

}
\seealso{
\code{\link{causalQual}}
}
\author{
Riccardo Di Francesco
}
